chore: rename to set_readonly_test
